Incorporates basic Redux structure
Adds an app-level heirarchy of components for the datastore, code-splitting,
persistence, and routing. Adds reducers and actions. Adds loading component.
Introduces Typescript evaluations. Adds Typescript types for packages. Modifies
code to satisfy Typescript rules. Updates code formatting rules.

This commit is as large as it is (And covers so many concerns) because it is a
port of existing .js code from a similar boilerplate.