Incorporates basic React Routing
Adds react router. Adds a header, footer, and body section to the application.
Modifies code to satisfy Typescript rules. Updates code formatting rules. Adds
Home and Error pages. Adds various constants and utils to support routing.

This commit is as large as it is (And covers so many concerns) because it is a
port of existing .js code from a similar boilerplate.